Clean up new menu and prompt structure, add documentation
BREAKING:

The way menu items for menu modules are configured now works
differently, expecting types from the [`menu`](##menu/menu) module.
The way commands declare themselves to be part of a menu group is also
different—the information previously stored in the `menuGroup` and
`display` properties now goes, in a somewhat different format, in the
[`menu`](##CommandSpec.menu) property.

The command parameter prompting functionality was changed. The
`paramHandler` option has been replaced by a
[`commandParamPrompt`](##commandParamPrompt) option. The prompting
functionality now lives in the [`prompt`](##ui/prompt) module, and
should be easier to extend.

The styling and animation of menus and tooltips was changed to be
simpler and easier to maintain. Fancy UI looks are now considered out
of scope for this module, and something that should be implemented in
third-party modules.

FEATURE:

New menu building blocks in the [`menu`](##menu/menu) module allow
more control and flexibility when defining menus.